+++ R/travis.R
@@ -1,0 +1,22 @@
+#' Add basic travis template to a package
+#'
+#' Also adds \code{.travis.yml} to \code{.Rbuildignore} so it isn't included
+#' in the built package
+#'
+#' @param pkg package description, can be path or package name.  See
+#'   \code{\link{as.package}} for more information
+add_travis <- function(pkg = ".") {
+  pkg <- as.package(pkg)
+
+  path <- file.path(pkg$path, ".travis.yml")
+  if (file.exists(path)) {
+    stop(".travis.yml already exists", call. = FALSE)
+  }
+
+  template_path <- system.file("templates/.travis.yml", package = "devtools")
+  file.copy(template_path, path)
+
+  add_build_ignore(pkg, ".travis.yml")
+
+  invisible(TRUE)
+}